# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exocortex Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assets/{asset_name}/blocks": {
      "post": {
        "tags": [
          "blocks"
        ],
        "summary": "Create Block",
        "operationId": "create_block_api_assets__asset_name__blocks_post",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/blocks/{block_id}": {
      "delete": {
        "tags": [
          "blocks"
        ],
        "summary": "Delete Block",
        "operationId": "delete_block_api_assets__asset_name__blocks__block_id__delete",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Block Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/blocks/selection": {
      "post": {
        "tags": [
          "blocks"
        ],
        "summary": "Update Selection",
        "operationId": "update_selection_api_assets__asset_name__blocks_selection_post",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSelectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/groups/merge": {
      "post": {
        "tags": [
          "blocks"
        ],
        "summary": "Merge Group",
        "operationId": "merge_group_api_assets__asset_name__groups_merge_post",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/groups/{group_idx}": {
      "delete": {
        "tags": [
          "blocks"
        ],
        "summary": "Delete Group",
        "operationId": "delete_group_api_assets__asset_name__groups__group_idx__delete",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "group_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Idx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/ui-state": {
      "put": {
        "tags": [
          "blocks"
        ],
        "summary": "Update Ui State",
        "operationId": "update_ui_state_api_assets__asset_name__ui_state_put",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUiStateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/system/config": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get App Config",
        "operationId": "get_app_config_api_system_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfigModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Update App Config",
        "operationId": "update_app_config_api_system_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfigModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppConfigModel": {
        "properties": {
          "themeMode": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "title": "Thememode",
            "default": "light"
          },
          "sidebarTextLineClamp": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Sidebartextlineclamp",
            "default": 1
          },
          "sidebarFontSizePx": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 10.0,
            "title": "Sidebarfontsizepx",
            "default": 14
          },
          "tutorReasoningEffort": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "xhigh"
            ],
            "title": "Tutorreasoningeffort",
            "default": "medium"
          },
          "tutorWithGlobalContext": {
            "type": "boolean",
            "title": "Tutorwithglobalcontext",
            "default": true
          }
        },
        "type": "object",
        "title": "AppConfigModel"
      },
...
      "AssetStateAssetModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pageCount": {
            "type": "integer",
            "title": "Pagecount"
          },
          "pdfPath": {
            "type": "string",
            "title": "Pdfpath"
          }
        },
        "type": "object",
        "required": [
          "name",
          "pageCount",
          "pdfPath"
        ],
        "title": "AssetStateAssetModel"
      },
      "AssetStateModel": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetStateAssetModel"
          },
          "references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "References"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockModel"
            },
            "type": "array",
            "title": "Blocks"
          },
          "mergeOrder": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Mergeorder"
          },
          "nextBlockId": {
            "type": "integer",
            "title": "Nextblockid",
            "default": 1
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupModel"
            },
            "type": "array",
            "title": "Groups"
          },
          "uiState": {
            "$ref": "#/components/schemas/UiStateModel"
          }
        },
        "type": "object",
        "required": [
          "asset"
        ],
        "title": "AssetStateModel"
      },
...
      "BlockModel": {
        "properties": {
          "blockId": {
            "type": "integer",
            "title": "Blockid"
          },
          "pageIndex": {
            "type": "integer",
            "title": "Pageindex"
          },
          "fractionRect": {
            "$ref": "#/components/schemas/RectModel"
          },
          "groupIdx": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupidx"
          }
        },
        "type": "object",
        "required": [
          "blockId",
          "pageIndex",
          "fractionRect"
        ],
        "title": "BlockModel"
      },
...
      "CreateBlockRequest": {
        "properties": {
          "pageIndex": {
            "type": "integer",
            "title": "Pageindex"
          },
          "fractionRect": {
            "$ref": "#/components/schemas/RectModel"
          }
        },
        "type": "object",
        "required": [
          "pageIndex",
          "fractionRect"
        ],
        "title": "CreateBlockRequest"
      },
...
      "GroupModel": {
        "properties": {
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "blockIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Blockids"
          }
        },
        "type": "object",
        "required": [
          "groupIdx"
        ],
        "title": "GroupModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MergeGroupRequest": {
        "properties": {
          "blockIds": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockids"
          },
          "markdownContent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdowncontent"
          },
          "groupIdx": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupidx"
          }
        },
        "type": "object",
        "title": "MergeGroupRequest"
      },
...
      "RectModel": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "RectModel"
      },
...
      "UiStateModel": {
        "properties": {
          "currentPage": {
            "type": "integer",
            "title": "Currentpage",
            "default": 1
          },
          "zoom": {
            "type": "number",
            "title": "Zoom",
            "default": 1.0
          },
          "pdfScrollFraction": {
            "type": "number",
            "title": "Pdfscrollfraction",
            "default": 0.0
          },
          "pdfScrollLeftFraction": {
            "type": "number",
            "title": "Pdfscrollleftfraction",
            "default": 0.0
          },
          "currentMarkdownPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentmarkdownpath"
          },
          "openMarkdownPaths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Openmarkdownpaths"
          },
          "sidebarCollapsed": {
            "type": "boolean",
            "title": "Sidebarcollapsed",
            "default": false
          },
          "sidebarCollapsedNodeIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sidebarcollapsednodeids"
          },
          "markdownScrollFractions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Markdownscrollfractions"
          },
          "sidebarWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarwidthratio"
          },
          "rightRailWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rightrailwidthratio"
          }
        },
        "type": "object",
        "title": "UiStateModel"
      },
      "UpdateAppConfigRequest": {
        "properties": {
          "themeMode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "light",
                  "dark"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Thememode"
          },
          "sidebarTextLineClamp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebartextlineclamp"
          },
          "sidebarFontSizePx": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 24.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarfontsizepx"
          },
          "tutorReasoningEffort": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutorreasoningeffort"
          },
          "tutorWithGlobalContext": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutorwithglobalcontext"
          }
        },
        "type": "object",
        "title": "UpdateAppConfigRequest"
      },
...
      "UpdateSelectionRequest": {
        "properties": {
          "mergeOrder": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Mergeorder"
          }
        },
        "type": "object",
        "title": "UpdateSelectionRequest"
      },
      "UpdateUiStateRequest": {
        "properties": {
          "currentPage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentpage"
          },
          "zoom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zoom"
          },
          "pdfScrollFraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdfscrollfraction"
          },
          "pdfScrollLeftFraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdfscrollleftfraction"
          },
          "currentMarkdownPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentmarkdownpath"
          },
          "openMarkdownPaths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openmarkdownpaths"
          },
          "sidebarCollapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarcollapsed"
          },
          "sidebarCollapsedNodeIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarcollapsednodeids"
          },
          "markdownScrollFractions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdownscrollfractions"
          },
          "sidebarWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarwidthratio"
          },
          "rightRailWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rightrailwidthratio"
          }
        },
        "type": "object",
        "title": "UpdateUiStateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}